🦄 refactor: 精简搜索建议中冗余代码
